feat: add quest context
