+++ src/AcceptedQuestList.jsx
@@ -1,67 +1,9 @@
-import React, { useState, useEffect } from 'react';
-import { supabase } from './supabaseClient';
+import React from 'react';
+import { useQuests } from './QuestContext.jsx';
 import './world.css';
 
 export default function AcceptedQuestList() {
-  const [quests, setQuests] = useState([]);
-  const [userId, setUserId] = useState(null);
-
-  useEffect(() => {
-    const stored = localStorage.getItem('quests');
-    setQuests(stored ? JSON.parse(stored) : []);
-    const handler = () => {
-      const updated = localStorage.getItem('quests');
-      setQuests(updated ? JSON.parse(updated) : []);
-    };
-    window.addEventListener('questsChange', handler);
-    return () => window.removeEventListener('questsChange', handler);
-  }, []);
-
-  useEffect(() => {
-    const load = async () => {
-      if (!navigator.onLine) return;
-      const {
-        data: { user },
-      } = await supabase.auth.getUser();
-      if (user) {
-        setUserId(user.id);
-        const { data } = await supabase
-          .from('quests')
-          .select('*')
-          .eq('user_id', user.id);
-        if (data) {
-          setQuests(data);
-          localStorage.setItem('quests', JSON.stringify(data));
-        }
-      }
-    };
-    load();
-  }, []);
-
-  const completeQuest = (id) => {
-    const all = quests.map((q) => {
-      if (q.id === id) {
-        const newResource = (parseInt(localStorage.getItem('resourceR') || '0', 10) + (q.resource || 0));
-        localStorage.setItem('resourceR', newResource);
-        if (userId && navigator.onLine) {
-          supabase.from('profiles').update({ resources: newResource }).eq('id', userId);
-        }
-        window.dispatchEvent(new CustomEvent('resourceChange', { detail: { resource: newResource } }));
-        return { ...q, completed: true };
-      }
-      return q;
-    });
-    setQuests(all);
-    localStorage.setItem('quests', JSON.stringify(all));
-    if (userId && navigator.onLine) {
-      supabase
-        .from('quests')
-        .update({ completed: true })
-        .eq('user_id', userId)
-        .eq('id', id);
-    }
-    window.dispatchEvent(new Event('questsChange'));
-  };
+  const { quests, completeQuest } = useQuests();
 
   return (
     <div>
